feat: tasks can build their status

+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/monitor/AutocraftingTaskButton.java
@@ -64,7 +64,9 @@ class AutocraftingTaskButton extends AbstractButton {
         final int textY = getY() + yOffset;
         text.render(graphics, textX, textY, Minecraft.getInstance().font, isHovered);
         final int ySpacing = SmallText.isSmall() ? 7 : 8;
-        final int percentageCompleted = Math.round(percentageCompletedProvider.getPercentageCompleted(task.id()) * 100);
+        final long percentageCompleted = Math.round(
+            percentageCompletedProvider.getPercentageCompleted(task.id()) * 100
+        );
         SmallText.render(graphics, Minecraft.getInstance().font, percentageCompleted + "%", textX, textY + ySpacing,
             0xFFFFFF, true);
         updateTooltip();
@@ -115,6 +117,6 @@ class AutocraftingTaskButton extends AbstractButton {
     }
 
     interface PercentageCompletedProvider {
-        float getPercentageCompleted(TaskId taskId);
+        double getPercentageCompleted(TaskId taskId);
     }
 }
